Save system's essentials were established and some tile sets were remoeved to create again with better organizasitpn

+++ Assets/_Project/Scripts/GameManagement/PlayerData.cs
@@ -12,8 +12,8 @@ public class PlayerData
     // -------------------------------------------------------------------------
     // Temel Statlar
     // -------------------------------------------------------------------------
-    public int   currentHP;
-    public int   maxHP;
+    public float currentHP;
+public float maxHP;
     public float currentStamina;
     public float maxStamina;
